requirements.txt
ultralytics>=8.0.0
opencv-python-headless
onnxruntime>=1.20.0
pillow>=10.4.0
gradio>=5.6.0
fpdf2>=2.8.0
werkzeug>=3.0.3
sqlalchemy>=2.0.0
flask>=3.0.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
alembic>=1.13.0

# Number Plate Detection (ALPR)
fast-alpr>=0.1.0
open-image-models>=0.1.0
fast-plate-ocr>=0.1.0

# Testing dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-flask>=1.3.0
pytest-mock>=3.12.0

# API Documentation
flasgger>=0.9.7.1

# Error Monitoring
sentry-sdk>=1.40.0

# Analytics & Data Export
pandas>=2.0.0
openpyxl>=3.1.0

# Rate Limiting
flask-limiter>=3.5.0

# Caching
redis>=5.0.0
flask-caching>=2.1.0

# Email
flask-mail>=0.9.1

# WebSocket Support
flask-socketio>=5.3.0
